move all components to separate folders
