+++ src/components/graphics/Graphic.js
